AAE-42641 testing with the placement change
